refactor(Race): SRP준수를 위해 Race 클래스 분리
Game:  게임의 전체 흐름을 관리 (입력, 레이스 실행, 결과 출력)
Race: 레이스의 실행 및 자동차 이동

이라는 각각의 일을 하게 함
